+++ ckanext/datastore/helpers.py
@@ -67,22 +67,27 @@ def should_fts_index_field_type(field_type):
     return field_type.lower() in ['tsvector', 'text', 'number']
 
 
-def get_table_names_from_sql(context, sql):
-    '''Parses the output of EXPLAIN (FORMAT JSON) looking for table names
+def get_table_and_function_names_from_sql(context, sql):
+    '''Parses the output of EXPLAIN (FORMAT JSON) looking for table and
+    function names
 
     It performs an EXPLAIN query against the provided SQL, and parses
-    the output recusively looking for "Relation Name".
+    the output recusively.
 
     Note that this requires Postgres 9.x.
 
     :param context: a CKAN context dict. It must contain a 'connection' key
         with the current DB connection.
     :type context: dict
-    :param sql: the SQL statement to parse for table names
+    :param sql: the SQL statement to parse for table and function names
     :type sql: string
 
-    :rtype: list of strings
+    :rtype: a tuple with two list of strings, one for table and one for
+    function names
     '''
+
+    table_names = []
+    function_names = []
 
     def _get_table_names_from_plan(plan):
 
@@ -99,13 +104,16 @@ def get_table_names_from_sql(context, sql):
 
         return table_names
 
+    function_names.extend(_get_function_names_from_sql(sql))
+
     result = context['connection'].execute(
         'EXPLAIN (FORMAT JSON) {0}'.format(sql.encode('utf-8'))).fetchone()
 
-    table_names = []
-
     try:
-        query_plan = json.loads(result['QUERY PLAN'])
+        if isinstance(result['QUERY PLAN'], basestring):
+            query_plan = json.loads(result['QUERY PLAN'])
+        else:
+            query_plan = result['QUERY PLAN']
         plan = query_plan[0]['Plan']
 
         table_names.extend(_get_table_names_from_plan(plan))
@@ -113,7 +121,25 @@ def get_table_names_from_sql(context, sql):
     except ValueError:
         log.error('Could not parse query plan')
 
-    return table_names
+    return table_names, function_names
+
+
+def _get_function_names_from_sql(sql):
+    function_names = []
+
+    def _get_function_names(tokens):
+        for token in tokens:
+            if isinstance(token, sqlparse.sql.Function):
+                function_name = token.get_name()
+                if function_name not in function_names:
+                    function_names.append(function_name)
+            if hasattr(token, 'tokens'):
+                _get_function_names(token.tokens)
+
+    parsed = sqlparse.parse(sql)[0]
+    _get_function_names(parsed.tokens)
+
+    return function_names
 
 
 def datastore_dictionary(resource_id):